Create arcs.common, Id (stub), Literal, and move Referencable to it.
