+++ src_kt/java/arcs/common/BUILD
@@ -1,0 +1,8 @@
+package(default_visibility = ["//visibility:public"])
+
+load("//build_defs:build_defs.bzl", "kt_arcs_library")
+
+kt_arcs_library(
+    name = "common",
+    srcs = glob(["*.kt"]),
+)
