+++ src_kt/java/arcs/common/Literal.kt
@@ -9,8 +9,9 @@
  * http://polymer.github.io/PATENTS.txt
  */
 
-package arcs.crdt.internal
+package arcs.common
 
-/** Denotes an individual actor responsible for modifications to a Crdt. */
-typealias Actor = String
+interface Literal
 
+/** A list of [Literal]s that is itself a [Literal]. */
+class LiteralList<T : Literal>(private val items: List<T>) : List<T> by items, Literal
