+++ src_kt/java/arcs/crdt/CrdtSingleton.kt
@@ -14,8 +14,8 @@ package arcs.crdt
 import arcs.crdt.CrdtSet.Operation.Add
 import arcs.crdt.CrdtSet.Operation.Remove
 import arcs.crdt.internal.Actor
-import arcs.crdt.internal.Referencable
-import arcs.crdt.internal.ReferenceId
+import arcs.common.Referencable
+import arcs.common.ReferenceId
 import arcs.crdt.internal.VersionMap
 
 /** A [CrdtModel] capable of managing a mutable reference. */
